[rocjitsu] correct RELATIVE64 ABI value, close no-metadata fail-open, fix sign-compare
  Address re-review findings on the text-relocation branch.

  Relocation guard (amdgpu_elf.h, tests/dbt/translate_test.cpp):
  - R_AMDGPU_RELATIVE64 was defined as 10, but the AMDGPU ELF ABI value is
  13
    (10 is R_AMDGPU_REL32_LO). The symbol-zero RELATIVE64 text-addend guard
  in
    has_relocation_to_text_symbol() therefore never matched a real record,
  so
    the fail-open it was meant to close stayed open. The tests masked it by
    hardcoding the same wrong 10. Correct the constant to 13, have both
  tests
    consume the production constant, and add a static_assert pinning it to
  the
    ABI value so a future drift cannot silently re-mask the bug.

  Virtual-LDS dispatch (rj_hsa_dbt_hooks.cpp):
  - Close the remaining fail-open: a kernel with no virtual-LDS variant
  whose
    packet group segment exceeds the host LDS limit returned KeepNormal and
    submitted an oversized (host-faulting) launch. There is no sidecar to
  fall
    back to, so this now calls abort_virtual_lds_dispatch (kernel
    "<unregistered>") like the post-plan paths. Updated the prepare_ doc so
    nullopt strictly means "normal descriptor is safe".

  Build (cdna4_to_cdna3_virtual_lds.cpp):
  - Fix -Werror=sign-compare in append_cdna3_virtual_lds_entry_prologue: the
    uint16_t base_sgpr/prologue_temp_sgpr + 1 promotes to int and was
  compared
    against the uint32_t kCdnaOrdinarySgprLimit. Cast both to uint32_t.

+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/code/dbt/semantic/cdna4_to_cdna3_virtual_lds.cpp
@@ -678,8 +678,10 @@ reserve_cdna3_virtual_lds_base_sgpr_pair(TranslationContext &context, KernelBloc
     return false;
   if ((translation.virtual_lds_lowering.prologue_temp_sgpr % 2) != 0)
     return false;
-  if (translation.virtual_lds_lowering.base_sgpr + 1 >= kCdnaOrdinarySgprLimit ||
-      translation.virtual_lds_lowering.prologue_temp_sgpr + 1 >= kCdnaOrdinarySgprLimit ||
+  if (static_cast<uint32_t>(translation.virtual_lds_lowering.base_sgpr) + 1 >=
+          kCdnaOrdinarySgprLimit ||
+      static_cast<uint32_t>(translation.virtual_lds_lowering.prologue_temp_sgpr) + 1 >=
+          kCdnaOrdinarySgprLimit ||
       pointer_base_sgpr > 126)
     return false;
   if (translation.lds_overflow_kernarg_pointer_offset > kCdnaSmemImmediateByteOffsetMax)
